Dissertation research paper:
Predicting prices in stock and currency markets using Machine Learning algorithms 

Abstract — This paper delves into the integration of Machine Learning (ML) algorithms within a trading software to optimize trading performance. It begins by critically examining the theoretical framework that steered this research, highlighting the existing work in the field as well as the absence of an end-user application. Consequently, it presents a revolutionary trading software architecture, meticulously crafted by blending ML algorithms, classic indicators and sentiment analysis. Notably, the system employs a Bidirectional LSTM model in a sequence-to-sequence configuration, predicting a week's worth of stock prices based on historical data. This approach is augmented by the capability to fetch real-time stock data and allows users to select specific features for modeling. The accuracy of the model is evaluated using metrics such as RMSE, MAPE, and r2 score, providing a comprehensive view of its performance.  In the subsequent sections, an in-depth analysis of the technical components of the software is provided, emphasizing the more novel characteristics that enhance trading efficiency. The paper will highlight the process of experimenting with different parameters to find the most suitable conditions to train the model. Ultimately, the paper contends that while the integration of ML in trading software can accurately predict stock prices and thus boost trading performance, the practical application and real-world effectiveness of such systems necessitate continuous refinement based on extensive testing, user feedback, and market dynamics. The efficacy of this project lies not just in the sophistication of its algorithms but also in its ability to visually interpret confluence between classic trading indicators.